Handle bugs in Admin APIs

+++ functions/src/common/models/Admin/subAdmin.ts
@@ -24,13 +24,21 @@ export const subAdminList = async (req: any, res: any, next: any) => {
       .limit(limit)
       .get();
 
-    const subAdminList = databaseQuery.docs.map((doc) => {
-      const data = doc.data();
-      return {
-        id: doc.id,
-        ...data,
-      };
-    });
+      console.log("databaseQuery ...........",databaseQuery)
+      const subAdminList = databaseQuery.docs.map((doc) => {
+        const data = doc.data();
+        return {
+          id: doc.id,
+          ...data,
+        };
+      });
+      if(!subAdminList.length){
+        return  res.status(404).send({
+          status: true,
+          message: "Admin not found.",
+          result: null,
+        });
+      }
     res.status(200).send({
       status: true,
       message: "Sub-admins fetched successfully",
@@ -57,6 +65,13 @@ export const updateStatus = async (req: any, res: any, next: any) => {
       .get();
 
     const getSubAdminData: any = databaseQuery.data();
+    if(!getSubAdminData){
+      return  res.status(404).send({
+        status: true,
+        message: "subadmin not found",
+        result: null
+      });
+    }
     getSubAdminData.status = status;
 
     const statusUpdate = await admin
@@ -89,6 +104,13 @@ export const deleteSubAdmin = async (req: any, res: any, next: any) => {
   try {
     const { subAdminId } = req.params;
     const subAdminRef = admin.firestore().collection("admin").doc(subAdminId);
+    if(!subAdminRef){
+      return  res.status(404).send({
+        status: true,
+        message: "subadmin not found",
+        result: null
+      });
+    }
     await subAdminRef.delete();
     res.status(200).send({
       status: true,
